call libc kill to end designated processes

+++ src/process/process_info.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub struct ProcessInfo {
-    pub pid: u64,
+    pub pid: i32,
     pub name: String,
     pub been_running_for_seconds: u64,
 }
